Workaround for RNScreens issue
